Refactoring: moved NodeParser classes to bl_nodes. Removed unused code.
