--- materialx/bl_nodes/nodes/__init__.py
@@ -1,10 +1,0 @@
-# SPDX-License-Identifier: GPL-2.0-or-later
-# Copyright 2022, AMD
-
-from .input import *
-from .output import *
-from .shader import *
-from .texture import *
-from .color import *
-from .converter import *
-from .vector import *
