+++ materialx/bl_nodes/vector.py
@@ -1,10 +1,9 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 # Copyright 2022, AMD
 
-from ..node_parser import NodeParser
+from .node_parser import NodeParser
+from . import log
 
-from ... import logging
-log = logging.Log("bl_nodes.nodes.vector")
 
 DEFAULT_SPACE = 'OBJECT'
